==> code/review analysis code/Step1_get_data_set_with_useful_features.ipynb <==
# Step 1: get data set with useful features

# load raw review data
import pandas as pd
review=pd.read_csv("all_review.csv")
review.head()

# remove useless columns
review=review.drop(columns=['Unnamed: 0','review_id','user_id','useful','funny','cool','date'])
review.head()

# save new data set
review.to_csv("review_with_useful_features.csv",index = False)
