==> code/review analysis code/step2_classifiy_adjectives_in_reviews_as_positive_or_negative.ipynb <==
# step 2: classifiy adjectives in reviews as positive or negative

import librarys and prepare data set

# librarys
import pandas as pd
import nltk
import sklearn
import nltk.sentiment
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.model_selection import train_test_split
from sklearn.naive_bayes import MultinomialNB
from sklearn.metrics import confusion_matrix

import re
import pandas as pd
import nltk
import sklearn
from nltk.stem import WordNetLemmatizer
import numpy as np
from nltk.tokenize import word_tokenize
import nltk.sentiment
import string
# nltk.download('stopwords')
# nltk.download('punkt')
# nltk.download('wordnet')
# nltk.download('averaged_perceptron_tagger')
from nltk.corpus import stopwords
from tqdm import tqdm
from sklearn import linear_model
import collections
from gensim.models import Word2Vec
from gensim.models import word2vec
import multiprocessing
import time
import inflect
from textblob import TextBlob
import keras.preprocessing.text as T
from keras.preprocessing.text import Tokenizer
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfTransformer

## define stopwords
sr = stopwords.words('english')
p = inflect.engine()
wnl = WordNetLemmatizer() 
table = str.maketrans('', '', string.punctuation)

# function to change text to words
def clean_review(text):
    #Change n't into not
    x = re.sub(r'n\'t',' not',text)
    #Change not word into not_word
    x = re.sub(r'not ','not_',x)
    #Split into words
    x = word_tokenize(x)
    #Remove punctuation
    x = [w.translate(table) if not re.match(r'not_.*', w) else w for w in x]
    #Change numbers into words
    x = [p.number_to_words(w) if w.isdigit() else w  for w in x ]
    #Remove non alphabetic
    x = [w for w in x if w.isalpha() or re.match(r'not_.*',w)]
    #Convert to lower case
    x = [w.lower() for w in x]
    #Remove stop words
    x = [w for w in x if not w in sr]
    ## lemmatization
    x = [wnl.lemmatize(w) for w in x]
    return x

# load data
review=pd.read_csv("review_with_useful_features.csv")
review.head()

# prepare data
review=review.drop(columns=['business_id'])
# we define stars from 1 to 3 as negative and stars from 4 to 5 as positive
review['stars']=np.where(review['stars'] >= 4, "positive", "negative")
review.head()

extract adjectives from text

# do word tokenization
clean_text = [[1]]*len(review)
for i in range(len(review)):
    clean_text[i]=clean_review(review['text'][i])

# get adjectives from all words
adj_words = [[1]]*len(review)
for i in range(len(review)):
    word=clean_text[i]
    pos_tag = nltk.pos_tag(word)
    adjs = []
    for words, pos in pos_tag:
        if (pos == 'JJ'):
            adjs.append(words)
    adj_words[i]=adjs

count frequencies for adjectives

# get all adjectives in one list
alladjwords = []
for i in range(len(review)):
    alladjwords=alladjwords + adj_words[i]
len(alladjwords)

# count the frequency for each adjectives
frequency = collections.Counter(alladjwords)
dict_frequency = dict(frequency)
print('the total number of adjectives in all reviews is',len(dict_frequency))

# get 1200 most common adjectives and then remove some wrong words
removeword=['u','dish','table','shrimp','noodle','sushi','salad','give','seafood','oyster','waitress','bread']
common = frequency.most_common(1200)
common=dict(common)
for word in removeword:
    common.pop(word)
print(common.keys())

get frequency matrix (row is the index of reviews and column is each adjective)

# get frequency matrix for adjectives
corpus=[]
corpus=list(review['text'])
df = pd.DataFrame(data=corpus, columns=['sentences'])
vectorizer = CountVectorizer(vocabulary=list(common.keys()), min_df=0)
X = vectorizer.fit_transform(df['sentences'].values)
result = pd.DataFrame(data=X.toarray(), columns=vectorizer.get_feature_names())
result.head()

use multinomial Naive Bayes model to classify adjectives as positive or negative

# fit multinomial Naive Bayes model
Y=review['stars']
model = MultinomialNB()
model.fit(result, Y)

# define classifier function and give two examples
def classifier(adjective):
    return model.predict(vectorizer.transform([adjective]))
print(classifier('great'))
print(classifier('bad'))

get and save dictionary for adjectives with positive or negative tags

# get dictionary
dict_adj={}
for word in common.keys():
    dict_adj[word]=classifier(word)[0]

# save dictionary
f = open('dict_adj.txt','w')
f.write(str(dict_adj))
f.close()
