==> code/review analysis code/step4_gather_reviews_from_the_same_restaurant.ipynb <==
# step 4: gather reviews from the same restaurant

# load review data
import pandas as pd
review=pd.read_csv("review_with_useful_features.csv")
review.head()

# load business data
business=pd.read_csv("../seafood_business.csv")
business.head()

divide review data set with different restaurant

# get restaurant position with review index
position=[0]
bid=review['business_id'][0]
restaurant=[bid]
for i in range(len(review)):
    if (review['business_id'][i]!=bid):
        bid=review['business_id'][i]
        position.append(i)
        restaurant.append(bid)
print(position[:10])
print(restaurant[:10])

connect reviews from one restaurant together and then save them in a new dataframe

# get bind text for each restaurant
bind_text=[]
for i in range(len(position)):
    if (i < (len(position)-1)):
        bind_text.append(" connection ".join(list(review['text'][position[i]:position[i+1]])))
    else:
        bind_text.append(" connection ".join(list(review['text'][position[i]:len(review)])))

# build new dataframe
bind_text=pd.DataFrame(bind_text)
bind_review = pd.concat([business['business_id'],business['stars'],bind_text], axis=1)
bind_review.columns = ["business_id", "stars", "bind_text"]
bind_review.head()

# save new data set
bind_review.to_csv("review_bind.csv",index = False)
